+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ### New features
+`core` module:
+- Add `Pagination` class to work with pagination
 `core-mit` module
 - Add `DetailsConfiguration` for `WrappedException.kt` and `EncodableResult.kt` to customize exception details inclusion
 - Added `randomBase64()` method to generate random Base64 of the specified length